requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
Markdown==3.2.1
Pillow==7.1.2
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
